requirements.txt
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0
bitsandbytes>=0.43.0
sentencepiece>=0.1.99
datasets>=2.18.0
scikit-learn>=1.4.0
numpy>=1.26.0
scipy>=1.12.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0
wandb>=0.16.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0.1
rouge-score>=0.1.2
huggingface-hub>=0.21.0
umap-learn>=0.5.0
